backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.7.0

# NLP
transformers==4.41.0
torch==2.3.0
accelerate==0.30.0
datasets==2.19.0
sentence-transformers==2.7.0
spacy==3.7.4
scikit-learn==1.5.0
faiss-cpu==1.8.0
numpy==1.26.4

# Temporal
py-heideltime==1.1.0

# Audio
openai-whisper==20231117
sounddevice==0.4.7

# PDF (keep existing functionality)
reportlab==4.2.0

# Utils
python-multipart==0.0.9
aiofiles==23.2.1
